+++ src/components/activation-display.tsx
