+++ src/components/traits-display.tsx
